+++ Examples/UtiliKit-iOSExample/Active Label/ActiveLabelViewController.swift
@@ -8,11 +8,11 @@
 import UIKit
 
 class ActiveLabelViewController: UIViewController {
+
     @IBOutlet private var activeLabels: [ActiveLabel]!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         activeLabels.forEach { $0.text = nil } 
     }
 }